Removed some debug and trying something new
git-svn-id: http://calder.sdml.cs.kent.edu/svn/srcDiffPurged/trunk/srcdiff/trunk@12049 f70b1346-9fe7-0310-9bf3-95a0de396e6b

+++ src/shortest_edit_script.h
@@ -13,7 +13,7 @@
 // edit constants
 const int COMMON = -1;
 const int INSERT = 0;
-const int DELETE = 1;
+const int DELETE = (int)1;
 const int CHANGE = 2;
 
 // single edit script
